+++ src/main/java/cn/net/yzl/ehr/dto/StaffTrainDto.java
@@ -69,10 +69,10 @@ public class StaffTrainDto implements Serializable {
     private Integer adjustDepartPostLater;
 
     @ApiModelProperty("调整前薪资")
-    private Double adjustSalaryFront;
+    private Integer adjustSalaryFront;
 
     @ApiModelProperty("调整后薪资")
-    private Double adjustSalaryLater;
+    private Integer adjustSalaryLater;
 
     @ApiModelProperty("异动时间")
     @JsonFormat(pattern="yyyy-MM-dd")
